+++ HISTORY.rst
@@ -4,6 +4,8 @@ Changelog
 1.0.0 (unreleased)
 ++++++++++++++++++
 
+- Move ``fields._to_fixed`` to ``utils.decimal_to_fixed``.
+
 1.0.0-a (2014-10-19)
 ++++++++++++++++++++